pyproject.toml
[project]
name = "matsuyama-jihen"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Hiroki Mizukami", email = "6128022+piroyoung@users.noreply.github.com" }
]
requires-python = ">=3.13"
dependencies = [
    "aiohttp>=3.12.13",
    "azure-eventhub>=5.15.0",
    "azure-identity>=1.23.0",
    "pydantic>=2.11.7",
]

[project.scripts]
publisher = "matsuyama_jihen.publisher.__main__:main"
subscriber = "matsuyama_jihen.subscriber.__main__:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/matsuyama_jihen"]

[dependency-groups]
dev = [
    "ruff>=0.12.0",
]
